Create loopquestion.cpp
A C++ program that allows an end user to answer Yes or No to a simple "Yes" or "No" question. This program will re-ask the same question, as long as the answer is neither Yes nor No.

Since C++ is case sensitive, I have also set up the program to accept both upper and lower case letters of the first letter of the allowed answers, since they both basically signify the same thing.